+++ changelog.md
@@ -7,11 +7,11 @@
   - By default, both should have --cookies-from-browser firefox.
   - The subscription list default should also have the date filter.
   - Maybe you should ask during setup for a browser preference?
-- [ ] Config Changes
+- [x] Config Changes
   - [x] Updater
     - [x] Make a config updater that runs upon startup
     - [x] Append a config version variable to the config file
-  - [ ] Before an update, make a backup of the previous config file
+  - [x] Before an update, make a backup of the previous config file
   - [x] Change `"single-lists":` to `"single":`
   - [x] Nest both list types underneath a new `"download-jobs"` category
   - [x] Move the current daterange, cookies, and browser variables to global flags.
@@ -22,6 +22,7 @@
   - You can now set default command line options for all download jobs within the single tab and subscriptions tab.
 - Config Updater
   - If an update changes the config format (like this one, for example), your options will automatically be migrated to the new format.
+  - A backup copy of the original config will be made prior to the update.
 - 
 
 ### Changes: